+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
 - Package header printed even when all issues were suppressed
+- Parsing fails for extraneous npm packages in package-lock.json
 
 ## [5.7.0] - 2023-08-24